fix: рефакторинг структуры проекта
